// src/lib/generated/requisition/discussions-v1/discussions-v1.ts
const mutationKey = ['createDiscussionApiV1DiscussionsKaizen AdminsKaizen AdminIdPost'];
// ...
const mutationKey = ['deleteDiscussionApiV1DiscussionsDiscussionIdDelete'];
// ...
export const addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost = (
    requisitionId: string,
    commentCreate: CommentCreate,
 signal?: AbortSignal
) => {
      
      
      return requisitionRequest<CommentResponse>(
      {url: `/api/v1/discussions/requisitions/${requisitionId}/comments`, method: 'POST',
      headers: {'Content-Type': 'application/json', },
      data: commentCreate, signal
    },
      );
    }
  


export const getAddCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPostMutationOptions = <TError = HTTPValidationError,
    TContext = unknown>(options?: { mutation?:UseMutationOptions<Awaited<ReturnType<typeof addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost>>, TError,{requisitionId: string;data: CommentCreate}, TContext>, }
): UseMutationOptions<Awaited<ReturnType<typeof addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost>>, TError,{requisitionId: string;data: CommentCreate}, TContext> => {

const mutationKey = ['addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost'];
const {mutation: mutationOptions} = options ?
      options.mutation && 'mutationKey' in options.mutation && options.mutation.mutationKey ?
      options
      : {...options, mutation: {...options.mutation, mutationKey}}
      : {mutation: { mutationKey, }};

      


      const mutationFn: MutationFunction<Awaited<ReturnType<typeof addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost>>, {requisitionId: string;data: CommentCreate}> = (props) => {
          const {requisitionId,data} = props ?? {};

          return  addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost(requisitionId,data,)
        }

        


  return  { mutationFn, ...mutationOptions }}
// ...
export const useAddCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost = <TError = HTTPValidationError,
    TContext = unknown>(options?: { mutation?:UseMutationOptions<Awaited<ReturnType<typeof addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost>>, TError,{requisitionId: string;data: CommentCreate}, TContext>, }
 , queryClient?: QueryClient): UseMutationResult<
        Awaited<ReturnType<typeof addCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPost>>,
        TError,
        {requisitionId: string;data: CommentCreate},
        TContext
      > => {

      const mutationOptions = getAddCommentApiV1DiscussionsKaizen AdminsKaizen AdminIdCommentsPostMutationOptions(options);

      return useMutation(mutationOptions, queryClient);
    }